refactor: useCardNumber 관련 유틸함수를 하나의 파일로 병합
